+++ sdk/resourcemanager/datafactory/armdatafactory/models.go
@@ -4825,12 +4825,20 @@ func (a *AzurePostgreSQLLinkedService) GetLinkedService() *LinkedService {
 
 // AzurePostgreSQLLinkedServiceTypeProperties - Azure PostgreSQL linked service properties.
 type AzurePostgreSQLLinkedServiceTypeProperties struct {
+	// Indicates the azure cloud type of the service principle auth. Allowed values are AzurePublic, AzureChina, AzureUsGovernment,
+	// AzureGermany. Default value is the data factory regions’ cloud type. Type:
+	// string (or Expression with resultType string).
+	AzureCloudType any
+
 	// The time to wait (in seconds) while trying to execute a command before terminating the attempt and generating an error.
 	// Set to zero for infinity. Type: integer.
 	CommandTimeout any
 
 	// An ODBC connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
 	ConnectionString any
+
+	// The credential reference containing authentication information.
+	Credential *CredentialReference
 
 	// Database name for connection. Type: string.
 	Database any
@@ -4858,6 +4866,29 @@ type AzurePostgreSQLLinkedServiceTypeProperties struct {
 
 	// Server name for connection. Type: string.
 	Server any
+
+	// The service principal credential type to use in Server-To-Server authentication. 'ServicePrincipalKey' for key/secret,
+	// 'ServicePrincipalCert' for certificate. Type: string (or Expression with
+	// resultType string).
+	ServicePrincipalCredentialType any
+
+	// Specify the base64 encoded certificate of your application registered in Azure Active Directory. Type: string (or Expression
+	// with resultType string).
+	ServicePrincipalEmbeddedCert SecretBaseClassification
+
+	// Specify the password of your certificate if your certificate has a password and you are using AadServicePrincipal authentication.
+	// Type: string (or Expression with resultType string).
+	ServicePrincipalEmbeddedCertPassword SecretBaseClassification
+
+	// The ID of the service principal used to authenticate against Azure Database for PostgreSQL Flexible server. Type: string
+	// (or Expression with resultType string).
+	ServicePrincipalID any
+
+	// The key of the service principal used to authenticate against Azure Database for PostgreSQL Flexible server.
+	ServicePrincipalKey SecretBaseClassification
+
+	// The name or ID of the tenant to which the service principal belongs. Type: string (or Expression with resultType string).
+	Tenant any
 
 	// The time to wait (in seconds) while trying to establish a connection before terminating the attempt and generating an error.
 	// Type: integer.
@@ -14049,15 +14080,45 @@ func (g *GreenplumLinkedService) GetLinkedService() *LinkedService {
 
 // GreenplumLinkedServiceTypeProperties - Greenplum Database linked service properties.
 type GreenplumLinkedServiceTypeProperties struct {
+	// The authentication type to use. Type: string. Only used for V2.
+	AuthenticationType *GreenplumAuthenticationType
+
+	// The time to wait (in seconds) while trying to execute a command before terminating the attempt and generating an error.
+	// Set to zero for infinity. Type: integer. Only used for V2.
+	CommandTimeout any
+
 	// An ODBC connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
 	ConnectionString any
+
+	// The time to wait (in seconds) while trying to establish a connection before terminating the attempt and generating an error.
+	// Type: integer. Only used for V2.
+	ConnectionTimeout any
+
+	// Database name for connection. Type: string. Only used for V2.
+	Database any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// Host name for connection. Type: string. Only used for V2.
+	Host any
+
+	// The Azure key vault secret reference of password in connection string. Type: string. Only used for V2.
+	Password SecretBaseClassification
+
+	// The port for the connection. Type: integer. Only used for V2.
+	Port any
+
 	// The Azure key vault secret reference of password in connection string.
 	Pwd *AzureKeyVaultSecretReference
+
+	// SSL mode for connection. Type: integer. 0: disable, 1:allow, 2: prefer, 3: require, 4: verify-ca, 5: verify-full. Type:
+	// integer. Only used for V2.
+	SSLMode any
+
+	// Username for authentication. Type: string. Only used for V2.
+	Username any
 }
 
 // GreenplumSource - A copy activity Greenplum Database source.
@@ -27191,8 +27252,8 @@ type SapOdpLinkedServiceTypeProperties struct {
 	// resultType string).
 	SncLibraryPath any
 
-	// SNC activation indicator to access the SAP server where the table is located. Must be either 0 (off) or 1 (on). Type: string
-	// (or Expression with resultType string).
+	// SNC activation flag (Boolean) to access the SAP server where the table is located. Type: boolean (or Expression with resultType
+	// boolean).
 	SncMode any
 
 	// Initiator's SNC name to access the SAP server where the table is located. Type: string (or Expression with resultType string).
@@ -27641,8 +27702,8 @@ type SapTableLinkedServiceTypeProperties struct {
 	// resultType string).
 	SncLibraryPath any
 
-	// SNC activation indicator to access the SAP server where the table is located. Must be either 0 (off) or 1 (on). Type: string
-	// (or Expression with resultType string).
+	// SNC activation flag (Boolean) to access the SAP server where the table is located. Type: boolean (or Expression with resultType
+	// boolean).
 	SncMode any
 
 	// Initiator's SNC name to access the SAP server where the table is located. Type: string (or Expression with resultType string).
@@ -28021,6 +28082,11 @@ type ScriptActivityScriptBlock struct {
 type ScriptActivityTypeProperties struct {
 	// Log settings of script activity.
 	LogSettings *ScriptActivityTypePropertiesLogSettings
+
+	// Enable to retrieve result sets from multiple SQL statements and the number of rows affected by the DML statement. Supported
+	// connector: SnowflakeV2. Type: boolean (or Expression with resultType
+	// boolean).
+	ReturnMultistatementResult any
 
 	// ScriptBlock execution timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
 	ScriptBlockExecutionTimeout any